requirements.txt
Flask
Flask-WTF
cryptography
mysql-connector-python
